refactor: Extract schema init and query builder from database.rs (1065 LOC)
Split the monolithic database.rs into focused modules:
- schema.rs: DDL constants, CREATE TABLE/INDEX, version tracking (~100 LOC)
- query_builder.rs: Dynamic WHERE/ORDER/LIMIT construction (~115 LOC)
- database_tests.rs: All 21 unit tests (~370 LOC)
- database.rs: Database struct with CRUD ops and row conversion (~310 LOC)
