+++ crates/memory/src/database.rs
@@ -12,6 +12,8 @@
 // SQLITE_OPEN_FULL_MUTEX is used as defense-in-depth.
 // Lock poisoning is handled gracefully by recovering the inner guard.
 
+use crate::query_builder;
+use crate::schema;
 use crate::types::{MemoryEntry, MemoryQuery, MemoryScope, MemoryType};
 use anyhow::{Context, Result};
 use chrono::DateTime;
@@ -26,9 +28,6 @@ const MAX_CONTENT_LENGTH: usize = 1_048_576;
 const MAX_TITLE_LENGTH: usize = 1_000;
 /// Maximum allowed agent_id length (256 characters)
 const MAX_AGENT_ID_LENGTH: usize = 256;
-
-/// Database schema version for migrations
-const SCHEMA_VERSION: i32 = 1;
 
 /// Thread-safe database handle
 ///
@@ -95,8 +94,10 @@ impl Database {
         };
 
         // Initialize schema
-        db.initialize_schema()
-            .context("Failed to initialize database schema")?;
+        {
+            let conn = db.lock_conn()?;
+            schema::initialize_schema(&conn).context("Failed to initialize database schema")?;
+        }
 
         Ok(db)
     }
@@ -113,107 +114,6 @@ impl Database {
                 Ok(poisoned.into_inner())
             }
         }
-    }
-
-    /// Initialize database schema
-    ///
-    /// Uses BEGIN EXCLUSIVE to serialize concurrent schema initialization.
-    /// The version check runs INSIDE the exclusive transaction to prevent
-    /// a TOCTOU race where another process could initialize the schema
-    /// between the version read and the lock acquisition.
-    fn initialize_schema(&self) -> Result<()> {
-        let conn = self.lock_conn()?;
-
-        // Acquire exclusive lock FIRST, then check version inside the transaction
-        // to eliminate the TOCTOU race condition (issue #378).
-        conn.execute_batch("BEGIN EXCLUSIVE")?;
-
-        // Check if schema exists (now safely inside the exclusive transaction)
-        let version: i32 = conn
-            .query_row(
-                "SELECT version FROM schema_version ORDER BY version DESC LIMIT 1",
-                [],
-                |row| row.get(0),
-            )
-            .unwrap_or(0);
-
-        if version >= SCHEMA_VERSION {
-            debug!("Schema version {} is current", version);
-            conn.execute_batch("COMMIT")?;
-            return Ok(());
-        }
-
-        info!("Initializing database schema version {}", SCHEMA_VERSION);
-
-        // Create schema_version table
-        conn.execute(
-            "CREATE TABLE IF NOT EXISTS schema_version (
-                version INTEGER PRIMARY KEY,
-                applied_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-            )",
-            [],
-        )
-        .context("Failed to create schema_version table")?;
-
-        // Create main memory_entries table
-        conn.execute(
-            "CREATE TABLE IF NOT EXISTS memory_entries (
-                id TEXT PRIMARY KEY,
-                agent_id TEXT NOT NULL,
-                session_id TEXT,
-                memory_type TEXT NOT NULL,
-                scope TEXT NOT NULL,
-                title TEXT NOT NULL,
-                content TEXT NOT NULL,
-                importance INTEGER NOT NULL DEFAULT 5,
-                tags TEXT, -- JSON array
-                metadata TEXT, -- JSON object
-                parent_id TEXT,
-                created_at TEXT NOT NULL,
-                updated_at TEXT NOT NULL,
-                expires_at TEXT,
-                FOREIGN KEY (parent_id) REFERENCES memory_entries(id) ON DELETE SET NULL
-            )",
-            [],
-        )
-        .context("Failed to create memory_entries table")?;
-
-        // Create indexes for fast queries
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_agent_id ON memory_entries(agent_id)",
-            [],
-        )?;
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_session_id ON memory_entries(session_id)",
-            [],
-        )?;
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_memory_type ON memory_entries(memory_type)",
-            [],
-        )?;
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_scope ON memory_entries(scope)",
-            [],
-        )?;
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_importance ON memory_entries(importance)",
-            [],
-        )?;
-        conn.execute(
-            "CREATE INDEX IF NOT EXISTS idx_created_at ON memory_entries(created_at)",
-            [],
-        )?;
-
-        // Record schema version (INSERT OR IGNORE to handle concurrent init)
-        conn.execute(
-            "INSERT OR IGNORE INTO schema_version (version) VALUES (?1)",
-            params![SCHEMA_VERSION],
-        )?;
-
-        conn.execute_batch("COMMIT")?;
-
-        info!("Database schema initialized successfully");
-        Ok(())
     }
 
     /// Validate content length limits on a memory entry
@@ -333,104 +233,10 @@ impl Database {
         }
     }
 
-    /// Escape LIKE wildcard characters in a search string
-    fn escape_like(input: &str) -> String {
-        input
-            .replace('\\', "\\\\")
-            .replace('%', "\\%")
-            .replace('_', "\\_")
-    }
-
     /// Retrieve memory entries matching the query
     pub fn query(&self, query: &MemoryQuery) -> Result<Vec<MemoryEntry>> {
         let conn = self.lock_conn()?;
-
-        // Build SQL query dynamically based on filters
-        let mut sql = String::from("SELECT * FROM memory_entries WHERE 1=1");
-        let mut param_values: Vec<Box<dyn rusqlite::ToSql>> = Vec::new();
-
-        if let Some(agent_id) = &query.agent_id {
-            sql.push_str(" AND agent_id = ?");
-            param_values.push(Box::new(agent_id.clone()));
-        }
-
-        if let Some(session_id) = &query.session_id {
-            sql.push_str(" AND session_id = ?");
-            param_values.push(Box::new(session_id.clone()));
-        }
-
-        if let Some(memory_type) = &query.memory_type {
-            sql.push_str(" AND memory_type = ?");
-            param_values.push(Box::new(memory_type.as_str().to_string()));
-        }
-
-        if let Some(scope) = &query.scope {
-            sql.push_str(" AND scope = ?");
-            param_values.push(Box::new(scope.as_str().to_string()));
-        }
-
-        if let Some(min_importance) = query.min_importance {
-            sql.push_str(" AND importance >= ?");
-            param_values.push(Box::new(min_importance));
-        }
-
-        if let Some(search) = &query.search {
-            sql.push_str(" AND (title LIKE ? ESCAPE '\\' OR content LIKE ? ESCAPE '\\')");
-            let escaped = Self::escape_like(search);
-            let pattern = format!("%{}%", escaped);
-            param_values.push(Box::new(pattern.clone()));
-            param_values.push(Box::new(pattern));
-        }
-
-        if let Some(created_after) = &query.created_after {
-            sql.push_str(" AND created_at >= ?");
-            param_values.push(Box::new(created_after.to_rfc3339()));
-        }
-
-        if let Some(created_before) = &query.created_before {
-            sql.push_str(" AND created_at <= ?");
-            param_values.push(Box::new(created_before.to_rfc3339()));
-        }
-
-        // Tag filtering in SQL using json_each() to avoid loading all rows into memory
-        if !query.tags.is_empty() {
-            for tag in &query.tags {
-                sql.push_str(
-                    " AND EXISTS (SELECT 1 FROM json_each(memory_entries.tags) WHERE json_each.value = ?)",
-                );
-                param_values.push(Box::new(tag.clone()));
-            }
-        }
-
-        // Filter out expired entries
-        let now = chrono::Utc::now().to_rfc3339();
-        sql.push_str(" AND (expires_at IS NULL OR expires_at > ?)");
-        param_values.push(Box::new(now));
-
-        // Order by importance and creation time
-        sql.push_str(" ORDER BY importance DESC, created_at DESC");
-
-        // Use parameterized LIMIT/OFFSET instead of format! interpolation
-        if let Some(limit) = query.limit {
-            sql.push_str(" LIMIT ?");
-            param_values.push(Box::new(limit as i64));
-        }
-
-        if let Some(offset) = query.offset {
-            sql.push_str(" OFFSET ?");
-            param_values.push(Box::new(offset as i64));
-        }
-
-        // Execute query
-        let param_refs: Vec<&dyn rusqlite::ToSql> =
-            param_values.iter().map(|p| p.as_ref()).collect();
-        let mut stmt = conn.prepare(&sql)?;
-        let entries = stmt
-            .query_map(param_refs.as_slice(), Self::row_to_entry)?
-            .collect::<Result<Vec<_>, _>>()?;
-
-        debug!("Query returned {} entries", entries.len());
-        Ok(entries)
+        query_builder::execute_query(&conn, query, Self::row_to_entry)
     }
 
     /// Get a specific memory entry by ID
@@ -576,490 +382,5 @@ pub struct MemoryStats {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    fn create_test_db() -> (Database, TempDir) {
-        let temp_dir = TempDir::new().unwrap();
-        let db_path = temp_dir.path().join("test.db");
-        let db = Database::open(&db_path).unwrap();
-        (db, temp_dir)
-    }
-
-    #[test]
-    fn test_database_creation() {
-        let (_db, _temp_dir) = create_test_db();
-        // Database should be created successfully
-    }
-
-    #[test]
-    fn test_store_and_retrieve() {
-        let (db, _temp_dir) = create_test_db();
-
-        let entry = MemoryEntry::new(
-            "test_agent",
-            "Test Memory",
-            "Test content",
-            MemoryType::Decision,
-            MemoryScope::Project,
-        )
-        .with_importance(8);
-
-        let id = db.store(&entry).unwrap();
-        assert_eq!(id, entry.id);
-
-        let retrieved = db.get(&id).unwrap().unwrap();
-        assert_eq!(retrieved.title, "Test Memory");
-        assert_eq!(retrieved.importance, 8);
-    }
-
-    #[test]
-    fn test_update() {
-        let (db, _temp_dir) = create_test_db();
-
-        let mut entry = MemoryEntry::new(
-            "test_agent",
-            "Original Title",
-            "Original content",
-            MemoryType::Decision,
-            MemoryScope::Project,
-        )
-        .with_importance(5);
-
-        let id = db.store(&entry).unwrap();
-
-        // Modify and update
-        entry.title = "Updated Title".to_string();
-        entry.content = "Updated content".to_string();
-        entry.importance = 9;
-
-        let updated = db.update(&entry).unwrap();
-        assert!(updated);
-
-        let retrieved = db.get(&id).unwrap().unwrap();
-        assert_eq!(retrieved.title, "Updated Title");
-        assert_eq!(retrieved.content, "Updated content");
-        assert_eq!(retrieved.importance, 9);
-        // updated_at should be newer than created_at (or at least equal, since time resolution)
-        assert!(retrieved.updated_at >= retrieved.created_at);
-    }
-
-    #[test]
-    fn test_update_nonexistent_returns_false() {
-        let (db, _temp_dir) = create_test_db();
-
-        let entry = MemoryEntry::new(
-            "test_agent",
-            "Ghost",
-            "Does not exist",
-            MemoryType::Context,
-            MemoryScope::Local,
-        );
-
-        let updated = db.update(&entry).unwrap();
-        assert!(!updated);
-    }
-
-    #[test]
-    fn test_query_filtering() {
-        let (db, _temp_dir) = create_test_db();
-
-        // Store multiple entries
-        for i in 0..5 {
-            let entry = MemoryEntry::new(
-                "test_agent",
-                format!("Memory {}", i),
-                format!("Content {}", i),
-                MemoryType::Decision,
-                MemoryScope::Project,
-            )
-            .with_importance((i + 5) as u8);
-            db.store(&entry).unwrap();
-        }
-
-        // Query with importance filter
-        let query = MemoryQuery::new().min_importance(7);
-        let results = db.query(&query).unwrap();
-        assert_eq!(results.len(), 3); // Entries with importance 7, 8, 9
-
-        // Query with limit
-        let query = MemoryQuery::new().limit(2);
-        let results = db.query(&query).unwrap();
-        assert_eq!(results.len(), 2);
-    }
-
-    #[test]
-    fn test_search_escapes_like_wildcards() {
-        let (db, _temp_dir) = create_test_db();
-
-        // Store entries with special LIKE characters in content
-        let entry1 = MemoryEntry::new(
-            "test_agent",
-            "Percent Entry",
-            "This has 100% coverage",
-            MemoryType::Context,
-            MemoryScope::Local,
-        );
-        db.store(&entry1).unwrap();
-
-        let entry2 = MemoryEntry::new(
-            "test_agent",
-            "Underscore Entry",
-            "Use snake_case naming",
-            MemoryType::Context,
-            MemoryScope::Local,
-        );
-        db.store(&entry2).unwrap();
-
-        let entry3 = MemoryEntry::new(
-            "test_agent",
-            "Normal Entry",
-            "Just normal content here",
-            MemoryType::Context,
-            MemoryScope::Local,
-        );
-        db.store(&entry3).unwrap();
-
-        // Searching for literal "%" should only match the percent entry
-        let query = MemoryQuery::new().search("100%");
-        let results = db.query(&query).unwrap();
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].title, "Percent Entry");
-
-        // Searching for literal "_" should only match the underscore entry
-        let query = MemoryQuery::new().search("snake_case");
-        let results = db.query(&query).unwrap();
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].title, "Underscore Entry");
-    }
-
-    #[test]
-    fn test_tag_filtering_in_sql() {
-        let (db, _temp_dir) = create_test_db();
-
-        let entry1 = MemoryEntry::new(
-            "test_agent",
-            "Tagged Entry",
-            "Content",
-            MemoryType::Decision,
-            MemoryScope::Project,
-        )
-        .add_tag("rust")
-        .add_tag("architecture");
-        db.store(&entry1).unwrap();
-
-        let entry2 = MemoryEntry::new(
-            "test_agent",
-            "Other Entry",
-            "Content",
-            MemoryType::Decision,
-            MemoryScope::Project,
-        )
-        .add_tag("python");
-        db.store(&entry2).unwrap();
-
-        // Query for "rust" tag
-        let query = MemoryQuery::new().add_tag("rust");
-        let results = db.query(&query).unwrap();
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].title, "Tagged Entry");
-
-        // Query for both "rust" and "architecture" tags
-        let query = MemoryQuery::new().add_tag("rust").add_tag("architecture");
-        let results = db.query(&query).unwrap();
-        assert_eq!(results.len(), 1);
-
-        // Query for a tag that doesn't match any entry
-        let query = MemoryQuery::new().add_tag("go");
-        let results = db.query(&query).unwrap();
-        assert_eq!(results.len(), 0);
-    }
-
-    #[test]
-    fn test_delete() {
-        let (db, _temp_dir) = create_test_db();
-
-        let entry = MemoryEntry::new(
-            "test_agent",
-            "To Delete",
-            "Content",
-            MemoryType::Context,
-            MemoryScope::Local,
-        );
-
-        let id = db.store(&entry).unwrap();
-        assert!(db.delete(&id).unwrap());
-        assert!(db.get(&id).unwrap().is_none());
-    }
-
-    #[test]
-    fn test_expiration_cleanup() {
-        let (db, _temp_dir) = create_test_db();
-
-        // Store expired entry
-        let expired = MemoryEntry::new(
-            "test_agent",
-            "Expired",
-            "Content",
-            MemoryType::Context,
-            MemoryScope::Local,
-        )
-        .with_expiration(chrono::Utc::now() - chrono::Duration::hours(1));
-
-        db.store(&expired).unwrap();
-
-        // Cleanup should remove it
-        let count = db.cleanup_expired().unwrap();
-        assert_eq!(count, 1);
-    }
-
-    #[test]
-    fn test_stats() {
-        let (db, _temp_dir) = create_test_db();
-
-        let stats = db.stats().unwrap();
-        assert_eq!(stats.total_entries, 0);
-
-        let entry = MemoryEntry::new(
-            "test_agent",
-            "Test",
-            "Content",
-            MemoryType::Context,
-            MemoryScope::Local,
-        );
-        db.store(&entry).unwrap();
-
-        let stats = db.stats().unwrap();
-        assert_eq!(stats.total_entries, 1);
-        assert!(stats.database_size_bytes > 0);
-    }
-
-    #[test]
-    fn test_concurrent_access() {
-        let (db, _temp_dir) = create_test_db();
-
-        let handles: Vec<_> = (0..10)
-            .map(|i| {
-                let db = db.clone();
-                std::thread::spawn(move || {
-                    let entry = MemoryEntry::new(
-                        format!("agent_{}", i),
-                        format!("Thread {} Memory", i),
-                        format!("Content from thread {}", i),
-                        MemoryType::Context,
-                        MemoryScope::Local,
-                    );
-                    db.store(&entry).unwrap();
-                })
-            })
-            .collect();
-
-        for handle in handles {
-            handle.join().unwrap();
-        }
-
-        let stats = db.stats().unwrap();
-        assert_eq!(stats.total_entries, 10);
-    }
-
-    #[test]
-    fn test_corrupted_database_file_returns_error() {
-        let temp_dir = TempDir::new().unwrap();
-        let db_path = temp_dir.path().join("corrupt.db");
-        std::fs::write(&db_path, b"this is not a sqlite database").unwrap();
-
-        let result = Database::open(&db_path);
-        // Should return an error, not panic
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_empty_string_inputs() {
-        let (db, _temp_dir) = create_test_db();
-
-        // Empty title and content should succeed (no validation rejects empty)
-        let entry = MemoryEntry::new("", "", "", MemoryType::Context, MemoryScope::Local);
-        let result = db.store(&entry);
-        assert!(result.is_ok());
-
-        let id = result.unwrap();
-        let retrieved = db.get(&id).unwrap().unwrap();
-        assert_eq!(retrieved.title, "");
-        assert_eq!(retrieved.content, "");
-        assert_eq!(retrieved.agent_id, "");
-    }
-
-    #[test]
-    fn test_concurrent_reads_while_writing() {
-        use std::sync::Barrier;
-
-        let (db, _temp_dir) = create_test_db();
-
-        // Seed some data first
-        for i in 0..5 {
-            let entry = MemoryEntry::new(
-                "agent",
-                format!("Seed {}", i),
-                "Content",
-                MemoryType::Context,
-                MemoryScope::Local,
-            );
-            db.store(&entry).unwrap();
-        }
-
-        let barrier = Arc::new(Barrier::new(12)); // 10 readers + 2 writers
-
-        // Spawn 10 reader threads and 2 writer threads concurrently
-        let mut handles = Vec::new();
-
-        for i in 0..10 {
-            let db = db.clone();
-            let barrier = barrier.clone();
-            handles.push(std::thread::spawn(move || {
-                barrier.wait();
-                // Concurrent read
-                let query = MemoryQuery::new().limit(10);
-                let results = db.query(&query).unwrap();
-                assert!(!results.is_empty(), "Reader {} got no results", i);
-            }));
-        }
-
-        for i in 0..2 {
-            let db = db.clone();
-            let barrier = barrier.clone();
-            handles.push(std::thread::spawn(move || {
-                barrier.wait();
-                // Concurrent write
-                let entry = MemoryEntry::new(
-                    "writer",
-                    format!("Written {}", i),
-                    "New content",
-                    MemoryType::Decision,
-                    MemoryScope::Project,
-                );
-                db.store(&entry).unwrap();
-            }));
-        }
-
-        for handle in handles {
-            handle.join().unwrap();
-        }
-
-        // Verify all writes completed
-        let stats = db.stats().unwrap();
-        assert_eq!(stats.total_entries, 7); // 5 seeded + 2 written
-    }
-
-    #[test]
-    fn test_content_length_limit_enforced() {
-        let (db, _temp_dir) = create_test_db();
-
-        // Content exceeding 1MB should be rejected
-        let huge_content = "x".repeat(MAX_CONTENT_LENGTH + 1);
-        let entry = MemoryEntry::new(
-            "agent",
-            "Big Entry",
-            huge_content,
-            MemoryType::Context,
-            MemoryScope::Local,
-        );
-        let result = db.store(&entry);
-        assert!(result.is_err());
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("Content exceeds maximum length"));
-    }
-
-    #[test]
-    fn test_title_length_limit_enforced() {
-        let (db, _temp_dir) = create_test_db();
-
-        let huge_title = "t".repeat(MAX_TITLE_LENGTH + 1);
-        let entry = MemoryEntry::new(
-            "agent",
-            huge_title,
-            "Normal content",
-            MemoryType::Context,
-            MemoryScope::Local,
-        );
-        let result = db.store(&entry);
-        assert!(result.is_err());
-        assert!(result
-            .unwrap_err()
-            .to_string()
-            .contains("Title exceeds maximum length"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn test_database_file_permissions() {
-        use std::os::unix::fs::PermissionsExt;
-
-        let temp_dir = TempDir::new().unwrap();
-        let db_path = temp_dir.path().join("perms_test").join("test.db");
-        let _db = Database::open(&db_path).unwrap();
-
-        // Verify file permissions are 0600 (owner read/write only)
-        let file_perms = std::fs::metadata(&db_path).unwrap().permissions();
-        assert_eq!(
-            file_perms.mode() & 0o777,
-            0o600,
-            "Database file should have 0600 permissions"
-        );
-
-        // Verify parent directory permissions are 0700 (owner only)
-        let dir_perms = std::fs::metadata(db_path.parent().unwrap())
-            .unwrap()
-            .permissions();
-        assert_eq!(
-            dir_perms.mode() & 0o777,
-            0o700,
-            "Parent directory should have 0700 permissions"
-        );
-    }
-
-    #[test]
-    fn test_database_recovers_from_poisoned_lock() {
-        use std::panic;
-
-        let (db, _temp_dir) = create_test_db();
-        let db_clone = db.clone();
-
-        // Deliberately poison the mutex by panicking while holding the lock
-        let handle = std::thread::spawn(move || {
-            let _ = panic::catch_unwind(panic::AssertUnwindSafe(|| {
-                let _guard = db_clone.lock_conn().unwrap();
-                panic!("Intentional panic to poison the lock");
-            }));
-        });
-
-        let _ = handle.join();
-
-        // The mutex is now poisoned, but lock_conn() should recover
-        // by unwrapping the PoisonError and returning the inner guard
-        let entry = MemoryEntry::new(
-            "test_agent",
-            "After Poison",
-            "Content after recovering from poisoned lock",
-            MemoryType::Context,
-            MemoryScope::Local,
-        );
-
-        // This should succeed because lock_conn() recovers from poisoning
-        let result = db.store(&entry);
-        assert!(
-            result.is_ok(),
-            "Database should recover from poisoned lock and allow operations"
-        );
-
-        // Verify the entry was actually stored
-        let id = result.unwrap();
-        let retrieved = db.get(&id).unwrap();
-        assert!(
-            retrieved.is_some(),
-            "Entry should be retrievable after lock recovery"
-        );
-    }
-}
+#[path = "database_tests.rs"]
+mod tests;
